+++ admin/src/app/main/pages/create-update/CreateUpdate.js
@@ -17,49 +17,8 @@ import 'grapesjs-preset-webpage/dist/grapesjs-preset-webpage.min.js'
 import '../../scripts/ScriptStyle.css';
 import Helper from 'src/app/helper';
 import { selectUser } from 'app/store/userSlice';
-
-const fullscreenEnable = () => {
-    var fullScrFn;
-    if(document.querySelector('.gjs-mdl-dialog').requestFullscreen){
-        fullScrFn = document.querySelector('.gjs-mdl-dialog').requestFullscreen();
-    } else if(
-        document.querySelector('.gjs-mdl-dialog').webkitRequestFullscreen ||
-        typeof document.querySelector('.gjs-mdl-dialog').webkitRequestFullscreen === 'undefined'
-    ){
-        document.querySelector('.gjs-mdl-dialog').webkitRequestFullscreen(); 
-        document.querySelector('.gjs-mdl-dialog').classList.add('gjs-fullscreen-mode');
-    } else if(document.querySelector('.gjs-mdl-dialog').msRequestFullscreen){
-        fullScrFn = document.querySelector('.gjs-mdl-dialog').msRequestFullscreen(); 
-    } else if(document.querySelector('.gjs-mdl-dialog').mozRequestFullScreen ){
-        fullScrFn = document.querySelector('.gjs-mdl-dialog').mozRequestFullScreen(); 
-    }
-
-    if(fullScrFn) {
-        fullScrFn.then(r => {
-            document.querySelector('.gjs-mdl-dialog').classList.add('gjs-fullscreen-mode');
-        })
-    }
-}
-
-const exitfullscreenWindow = () => {
-    var exitFn;
-    if(document.exitFullscreen) {
-        exitFn = document.exitFullscreen();
-    } else if(document.mozCancelFullScreen) {
-        exitFn = document.mozCancelFullScreen();
-    } else if(document.webkitExitFullscreen) {
-        document.webkitExitFullscreen();
-        document.querySelector('.gjs-mdl-dialog').classList.remove('gjs-fullscreen-mode')
-    }
-    if(exitFn) {
-        exitFn.then(r => {
-            document.querySelector('.gjs-mdl-dialog').classList.remove('gjs-fullscreen-mode')
-        })
-    }
-
-    document.querySelector('.minimize').classList.add('hide-btn');
-    document.querySelector('.maximize').classList.remove('hide-btn');
-}
+import { customComponents, scriptBlockManager, customCodeEditor } from '../componentPlugins'
+import { customCheckboxTrait, memberTrait } from '../traitPlugins'
 
 const CreateUpdate = () => {
     const module = 'pages';
@@ -67,7 +26,9 @@ const CreateUpdate = () => {
     const dispatch = useDispatch();
     const user = useSelector(selectUser);
     const moduleId = useParams().moduleId;
-    const storageKey = (moduleId !== 'create' && !isNaN(moduleId)) ? `gjs-pages-${moduleId}` : `gjs-pages-new`;
+    // const storageKey = '';
+    // const storageKey = (moduleId !== 'create' && !isNaN(moduleId)) ? `gjs-pages-${moduleId}` : `gjs-pages-new`;
+    const [storageKey, setStorageKey] = useState('');
     const [loading, setLoading] = useState(false);
     const [openAlertDialog, setOpenAlertDialog] = useState(false);
     const [editor, setEditor] = useState({});
@@ -77,6 +38,10 @@ const CreateUpdate = () => {
     const [components, setComponents] = useState([]);
     const domain = `https://${user.loggedin_portal.domain}/`;
     const [expanded, setExpanded] = useState('panel1');
+    // const [ready, setReady] = useState(false);
+    // const [scriptJson, setScriptJson] = useState([]);
+    // const [scriptBlocks, setScriptBlocks] = useState([]);
+
 
     const handleChangeExpand = (panel) => (event, newExpanded) => {
         setExpanded(newExpanded ? panel : false);
@@ -97,20 +62,21 @@ const CreateUpdate = () => {
 
     useEffect(() => {
         const editor = grapesjs.init({
+            // autorender: false,
             container: '#gjs',
             allowScripts: 1,
             protectedCss: '',   // disabled default CSS
             height: '700px',
             width: '100%',
             style: '.txt-red{color: red}',
-            plugins: ["gjs-preset-webpage"],
+            plugins: ["gjs-preset-webpage", customCodeEditor, customComponents, scriptBlockManager, customCheckboxTrait, memberTrait],
             storageManager: {
                 id: 'gjs-',
                 type: 'local',
                 options: {
                     local: { key: storageKey }
                 },
-                autosave: true,
+                autosave: !false,
                 storeComponents: true,
                 storeStyles: true,
                 storeHtml: true,
@@ -145,198 +111,52 @@ const CreateUpdate = () => {
                         flexGrid: 1,
                     },
                     blocks: ['link-block', 'quote', 'text-basic'],
-                },
+                }
             },
             modal: {
                 backdrop: false
             },
         });
-        moduleId === 'create' ? getLayoutOptions(editor) : '';
-        setEditor(editor);
-
-        const pfx = editor.getConfig().stylePrefix
-        const modal = editor.Modal
-        const cmdm = editor.Commands
-        const htmlCodeViewer = editor.CodeManager.getViewer('CodeMirror').clone()
-        const cssCodeViewer = editor.CodeManager.getViewer('CodeMirror').clone()
-        const pnm = editor.Panels
-        const fullscrBtn = document.createElement('button');
-        fullscrBtn.setAttribute('title', 'Fullscreen')
-        fullscrBtn.setAttribute('class', 'grapes-modal-editor-fullscreen maximize');
-        fullscrBtn.innerHTML = '<i class="fa fa-window-maximize" aria-hidden="true"></i>';
-        const minimizeBtn = document.createElement('button');
-        minimizeBtn.setAttribute('title', 'Exit Fullscreen')
-        minimizeBtn.setAttribute('class', 'grapes-modal-editor-fullscreen hide-btn minimize');
-        minimizeBtn.innerHTML = '<i class="fa fa-window-minimize" aria-hidden="true"></i>';
-        const closeBtn = document.createElement('button');
-        closeBtn.setAttribute('title', 'Close')
-        closeBtn.setAttribute('class', 'grapes-modal-editor-close');
-        closeBtn.innerHTML = '<i class="fa fa-times" aria-hidden="true"></i>'
-        const rootContainer = document.createElement('div');
-        rootContainer.setAttribute('class', 'grapes-modal-editor-container');
-        const editorHeader =  document.createElement('div');
-        editorHeader.setAttribute('class', 'grapes-modal-editor-header');        
-        editorHeader.append(fullscrBtn);
-        editorHeader.append(minimizeBtn);
-        editorHeader.append(closeBtn);
-        const editorBody = document.createElement('div');
-        editorBody.setAttribute('class', 'grapes-modal-editor-body');
-        const editorFooter = document.createElement('div');
-        editorFooter.setAttribute('class', 'grapes-modal-editor-footer');
-        const btnEdit = document.createElement('button');
-        rootContainer.append(editorHeader)
-
-        const codeViewerOpt = {
-            readOnly: 0,
-            theme: 'hopscotch',
-            autoBeautify: true,
-            autoCloseTags: true,
-            autoCloseBrackets: true,
-            lineWrapping: true,
-            styleActiveLine: true,
-            smartIndent: true,
-            indentWithTabs: true
-        }
-
-        htmlCodeViewer.set({
-            codeName: 'htmlmixed',
-            ...codeViewerOpt
-        })
-
-        cssCodeViewer.set({
-            codeName: 'css',
-            ...codeViewerOpt
-        });
-
-        btnEdit.innerHTML = 'Save'
-        btnEdit.className = pfx + 'btn-prim ' + pfx + 'btn-import'
-        btnEdit.onclick = function () {
-            const html = htmlCodeViewer.editor.getValue()
-            const css = cssCodeViewer.editor.getValue()
-            editor.DomComponents.getWrapper().set('content', '')
-            // editor.CssComposer.clear();            
-            // const HTML_CSS = html.trim() + `<style>${css}</style>`
-            editor.setComponents(html.trim());
-            editor.setStyle(css);
-            modal.close();
-            if(document.fullscreenElement !== null || document.webkitFullscreenElement !== null) {
-                exitfullscreenWindow();
-            }
-        }
-
-        closeBtn.onclick = function () {
-            modal.close();
-            if(document.fullscreenElement !== null || document.webkitFullscreenElement !== null) {
-                exitfullscreenWindow(); 
-            }
-        }
-        //-- Edit Code popup Fullscreen ON|OFF
-        minimizeBtn.onclick = function() {
-            exitfullscreenWindow();
-        }
-        fullscrBtn.onclick = function() {
-            minimizeBtn.classList.remove('hide-btn')
-            fullscrBtn.classList.add('hide-btn');
-            fullscreenEnable();            
-        } 
-        document.addEventListener("fullscreenchange", function(e) {
-            if(
-                (
-                    typeof document.fullscreenElement !== 'undefined' &&
-                    document.fullscreenElement == null
-                ) || (
-                    typeof document.webkitFullscreenElement !== 'undefined' &&
-                    document.webkitFullscreenElement == null
-                )
-            ) {
-                document.querySelector('.gjs-mdl-dialog').classList.remove('gjs-fullscreen-mode');
-                minimizeBtn.classList.add('hide-btn')
-                fullscrBtn.classList.remove('hide-btn');
-            }
-        });
-        //-----
-
-        cmdm.add('edit-code', {
-            run: function (editor, sender) {
-                sender && sender.set('active', 0)
-                var htmlViewer = htmlCodeViewer.editor
-                var cssViewer = cssCodeViewer.editor
-                modal.setTitle('Edit code')
-                var InnerHtml = editor.getHtml()
-                var Css = editor.getCss();
-                if (!htmlViewer && !cssViewer) {
-                    const txtarea = editorTextAreaCreate(editorBody, 'HTML')
-                    const cssarea = editorTextAreaCreate(editorBody, 'CSS')
-
-                    editorFooter.append(btnEdit)
-                    rootContainer.append(editorBody)
-                    rootContainer.append(editorFooter)
-                    htmlCodeViewer.init(txtarea)
-                    cssCodeViewer.init(cssarea)
-                    htmlViewer = htmlCodeViewer.editor
-                    cssViewer = cssCodeViewer.editor
-                }
-                modal.setContent('')
-                modal.setContent(rootContainer)
-                htmlCodeViewer.setContent(InnerHtml)
-                cssCodeViewer.setContent(Css)
-                modal.open({ attributes: { class: 'custom-code-editor' } })
-                htmlViewer.refresh()
-                cssViewer.refresh()
-            }
-        })
-
-        // To enable JS in the editor
-        cmdm.add('allowScripts', {
-            run: function (editor) {
-                editor.getConfig().allowScripts = 1;
-            },
-            stop: function (editor) {
-                editor.getConfig().allowScripts = 0;
-            },
-        });
-
-        // Removed default read-only code editor btn from toolbar
-        pnm.removeButton("options", 'export-template');
-
-        pnm.addButton('options',
-            [
-                {
-                    id: 'edit',
-                    className: 'fa fa-code',
-                    command: 'edit-code',
-                    attributes: {
-                        title: 'Edit Code'
-                    }
-                }
-            ]
-        );
-        
-
-        editor.onReady(() => {
-            loadEditorData(editor);
-
-            // Collapsed all the blocks accordian by default
-            const categories = editor.BlockManager.getCategories();
-            categories.each(category => {
-                category.set('open', false).on('change:open', opened => {
-                    opened.get('open') && categories.each(category => {
-                        category !== opened && category.set('open', false)
-                    })
-                })
-            });
-            //------------ End ----------
-        });
-
-        editor.on('change:changesCount', (model) => {
-            const changes = model.get('changesCount');
-            if (changes) {
-                setChangeCount(changeCount => changeCount + 1)
-            }
-        });
+        setEditor(editor); 
     }, []);
 
-    const editorTextAreaCreate = (editorBody, title) => {
+    useEffect(()=>{
+        if(Object.keys(editor).length){
+            getLayoutOptions(editor);           
+            
+            editor.onReady(() => {    
+                // Collapsed all the blocks accordian by default
+                const categories = editor.BlockManager.getCategories();
+                categories.each(category => {
+                    category.set('open', false).on('change:open', opened => {
+                        opened.get('open') && categories.each(category => {
+                            category !== opened && category.set('open', false)
+                        })
+                    })
+                });
+                //------------ End ----------
+            })
+        }
+    }, [editor])
+
+    /*useEffect(()=>{
+        getLayoutOptions();
+        
+    }, []);*/
+
+    /*useEffect(()=>{
+        if(scriptJson.length){
+            //blockManagerForScripts(scriptBlocks, editor)
+            // if (moduleId !== 'create' && !isNaN(moduleId)) {
+            //     getSingleRecordById(moduleId, editor);
+            // }
+            
+            loadEditorData(editor);
+
+        }
+    }, [scriptJson]);*/
+
+    /*const editorTextAreaCreate = (editorBody, title) => {
         const container = document.createElement('div')
         const childContainer = document.createElement('div')
         const titleContainer = document.createElement('div')
@@ -352,14 +172,16 @@ const CreateUpdate = () => {
         container.appendChild(childContainer)
         editorBody.appendChild(container)
         return txtarea
-    }
+    }*/
 
     const loadEditorData = async (editor) => {
         if (moduleId !== 'create' && !isNaN(moduleId)) {
             getSingleRecordById(moduleId, editor);
         } else {
             const storageManager = editor.Storage;
-            const data = storageManager.load();
+            storageManager.getStorageOptions()['key'] = 'gjs-pages-new';
+            setStorageKey('gjs-pages-new')
+            const data = await storageManager.load();           
             editor.loadProjectData(data);
             setAllData(prevData => {
                 return {
@@ -407,6 +229,7 @@ const CreateUpdate = () => {
         }));
         dynamicErrorMsg('name', event.target.value.trim());
     }
+
     const createCustomComponentForEditor = (components_val, editor) => {
         components_val.map((val) => {
             editor.BlockManager.add(`block-${val.value}`, {
@@ -421,6 +244,30 @@ const CreateUpdate = () => {
         // Custom Component block accordian set collapsed by default
         editor.BlockManager.getCategories()._byId["Custom Component"].set("open", false)
     }
+
+    /*const blockManagerForScripts = (jsonData, editor) => {
+        jsonData.map((val) => {
+            editor.BlockManager.add(`block-${val.value}`, {
+                label: val.value,
+                category: 'Scripts',
+                attributes: {
+                    class: 'fa fa-code'
+                },
+                /*content: (val.config_json)
+                        ? '<div data-script="'+val.id+'" data-gjs-type="'+val.config_json+'"></div>'
+                        : '<div data-script="'+val.id+'"></div>' *
+                content: (val.config_json) ? {type: 'component_'+val.id} : '<div data-script="'+val.id+'"></div>'
+            });
+        })
+        // Custom Component block accordian set collapsed by default
+        // editor.BlockManager.getCategories()._byId["Custom Component"].set("open", false)
+
+        // setTimeout(()=>{
+        //     editor.render();
+        //     console.log('render')
+        // }, 2000)
+    }*/
+
     const getLayoutOptions = (editor) => {
         axios.get(`/${module}/add`)
             .then((response) => {
@@ -431,7 +278,25 @@ const CreateUpdate = () => {
                         layout_id: response.data.results.fields.layout_id.value
                     });
                     setComponents(response.data.results.fields.components.options);
+                    loadEditorData(editor);
+
+                    /*const jsonData = response.data.results.fields.scripts.options;
+                    const componentJson = [];
+                    jsonData.map((val) => {
+                        if(val.config_json){
+                            componentJson.push({
+                                type_name: 'component_'+val.id,
+                                name: val.value,
+                                traits: val.config_json
+                            });
+                        }                  
+                    });
+                    setScriptJson(componentJson);
+                    setScriptBlocks(jsonData)
+                    console.log(componentJson)
+
                     createCustomComponentForEditor(response.data.results.fields.components.options, editor);
+                    blockManagerForScripts(response.data.results.fields.scripts.options, editor);*/
                 } else {
                     dispatch(showMessage({ variant: 'error', message: response.data.results.message }))
                 }
@@ -507,8 +372,13 @@ const CreateUpdate = () => {
                         meta_code: record.meta_code,
                         auth_required: !!record.auth_required,
                     }));
-                    createCustomComponentForEditor(response.data.results.fields.components.options, editor);
+                    // createCustomComponentForEditor(response.data.results.fields.components.options, editor);
+
+                    const storageManager = editor.Storage;
+                    storageManager.getStorageOptions()['key'] = `gjs-pages-${moduleId}`;
+                    setStorageKey(`gjs-pages-${moduleId}`);
                     editor.loadProjectData(record.page_json);
+
 
                     //-- Set to chnage state value to 0 because edior values fetched from DB and not done any changes by the user actually.
                     setChangeCount(changeCount => changeCount - 1);
